+++ UI/InfoPanels.h
@@ -356,9 +356,9 @@ private:
     boost::shared_ptr<GG::Texture>                  m_bar_shading_texture;
 
     std::vector<std::pair<MeterType, MeterType> >   m_meter_types;      // list of <MeterType, MeterType> pairs, where .first is the "actual" meter value, and .second is the the "target" or "max" meter value.  Either may be INVALID_METER_TYPE, in which case nothing will be shown for that MeterType
-    std::vector<double>                             m_initial_values;   // initial value of .first MeterTypes at the start of this turn
-    std::vector<double>                             m_projected_values; // projected current value of .first MeterTypes for the start of next turn
-    std::vector<double>                             m_target_max_values;// current values of the .second MeterTypes in m_meter_types
+    std::vector<float>                              m_initial_values;   // initial value of .first MeterTypes at the start of this turn
+    std::vector<float>                              m_projected_values; // projected current value of .first MeterTypes for the start of next turn
+    std::vector<float>                              m_target_max_values;// current values of the .second MeterTypes in m_meter_types
 
     int                                             m_object_id;
